wip. move ChangePolicy into its own file
