+++ src/bin/dashboard_src/send_screen.rs
@@ -11,8 +11,8 @@ use neptune_cash::config_models::network::Network;
 use neptune_cash::models::blockchain::type_scripts::native_currency_amount::NativeCurrencyAmount;
 use neptune_cash::models::state::wallet::address::KeyType;
 use neptune_cash::models::state::wallet::address::ReceivingAddress;
+use neptune_cash::models::state::wallet::change_policy::ChangePolicy;
 use neptune_cash::models::state::wallet::utxo_notification::UtxoNotificationMedium;
-use neptune_cash::models::state::ChangePolicy;
 use neptune_cash::rpc_auth;
 use neptune_cash::rpc_server::RPCClient;
 use neptune_cash::tx_initiation::builder::tx_output_list_builder::OutputFormat;
