+++ src/bin/neptune-cli.rs
@@ -23,6 +23,7 @@ use neptune_cash::models::blockchain::block::block_selector::BlockSelector;
 use neptune_cash::models::blockchain::type_scripts::native_currency_amount::NativeCurrencyAmount;
 use neptune_cash::models::state::wallet::address::KeyType;
 use neptune_cash::models::state::wallet::address::ReceivingAddress;
+use neptune_cash::models::state::wallet::change_policy::ChangePolicy;
 use neptune_cash::models::state::wallet::coin_with_possible_timelock::CoinWithPossibleTimeLock;
 use neptune_cash::models::state::wallet::secret_key_material::SecretKeyMaterial;
 use neptune_cash::models::state::wallet::utxo_notification::PrivateNotificationData;
@@ -31,7 +32,6 @@ use neptune_cash::models::state::wallet::wallet_file::WalletFile;
 use neptune_cash::models::state::wallet::wallet_file::WalletFileContext;
 use neptune_cash::models::state::wallet::wallet_status::WalletStatus;
 use neptune_cash::models::state::wallet::wallet_status::WalletStatusExportFormat;
-use neptune_cash::models::state::ChangePolicy;
 use neptune_cash::rpc_auth;
 use neptune_cash::rpc_server::error::RpcError;
 use neptune_cash::rpc_server::RPCClient;
