+++ src/models/state/wallet/mod.rs
@@ -1,4 +1,5 @@
 pub mod address;
+pub mod change_policy;
 pub mod coin_with_possible_timelock;
 pub(crate) mod expected_utxo;
 pub(crate) mod incoming_utxo;
