+++ src/tx_initiation/builder/transaction_details_builder.rs
@@ -18,9 +18,9 @@ use crate::models::blockchain::transaction::utxo::Utxo;
 use crate::models::blockchain::type_scripts::native_currency_amount::NativeCurrencyAmount;
 use crate::models::proof_abstractions::timestamp::Timestamp;
 use crate::models::state::transaction_details::TransactionDetails;
-use crate::models::state::tx_creation_config::ChangePolicy;
 use crate::models::state::wallet::address::KeyType;
 use crate::models::state::wallet::address::SpendingKey;
+use crate::models::state::wallet::change_policy::ChangePolicy;
 use crate::models::state::wallet::transaction_input::TxInput;
 use crate::models::state::wallet::transaction_input::TxInputList;
 use crate::models::state::wallet::transaction_output::TxOutput;
