+++ src/tx_initiation/export.rs
@@ -13,7 +13,7 @@ pub use crate::models::proof_abstractions::timestamp::Timestamp;
 pub use crate::models::state::transaction_details::TransactionDetails;
 pub use crate::models::state::transaction_kernel_id::TransactionKernelId;
 pub use crate::models::state::tx_creation_artifacts::TxCreationArtifacts;
-pub use crate::models::state::tx_creation_config::ChangePolicy;
+pub use crate::models::state::wallet::change_policy::ChangePolicy;
 pub use crate::models::state::wallet::transaction_input::TxInput;
 pub use crate::models::state::wallet::transaction_input::TxInputList;
 pub use crate::models::state::wallet::transaction_output::TxOutputList;
